Refactor register functions into utils
