+++ client/src/Utils/filterQuestionsUtil.js
@@ -1,0 +1,10 @@
+const filterQuestionsUtil = (array, string) => {
+  return array.filter(e => {
+    if (e.section.includes(string)) {
+      return e
+    }
+    return ""
+  })
+}
+
+export default filterQuestionsUtil
